+++ Libraries/LibWeb/HTML/Canvas/CanvasSettings.idl
@@ -1,0 +1,5 @@
+// https://html.spec.whatwg.org/multipage/canvas.html#canvassettings
+interface mixin CanvasSettings {
+    // settings
+    CanvasRenderingContext2DSettings getContextAttributes();
+};
